Remove WeakReferences (not necessary with the new design), reorganize code where it should belong (streams structure to ConnectionStreams, server time to HttpClient).
