+++ commons/src/main/java/com/pryv/connection/ConnectionEvents.java
@@ -1,23 +1,19 @@
 package com.pryv.connection;
 
-import com.pryv.AbstractConnection;
 import com.pryv.Filter;
 import com.pryv.api.HttpClient;
 import com.pryv.model.Event;
 import com.pryv.utils.JsonConverter;
 
 import java.io.IOException;
-import java.lang.ref.WeakReference;
 import java.util.List;
 
 public class ConnectionEvents {
 
-    private WeakReference<AbstractConnection> weakConnection;
     private static final String PATH = "events";
     private HttpClient httpClient;
 
-    public ConnectionEvents(WeakReference<AbstractConnection> weakConnection, HttpClient client) {
-        this.weakConnection = weakConnection;
+    public ConnectionEvents(HttpClient client) {
         this.httpClient = client;
     }
 
@@ -25,7 +21,6 @@ public class ConnectionEvents {
         HttpClient.ApiResponse apiResponse = httpClient.getRequest(PATH, filter).exec();
         List<Event> receivedEvents = JsonConverter.createEventsFromJson(apiResponse.getJsonBody());
         for (Event receivedEvent : receivedEvents) {
-            receivedEvent.assignConnection(weakConnection);
             Event.createOrReuse(receivedEvent);
         }
         // TODO: retrieve eventDeletions
@@ -46,7 +41,6 @@ public class ConnectionEvents {
 
         String json = apiResponse.getJsonBody();
         Event createdEvent = JsonConverter.retrieveEventFromJson(json);
-        createdEvent.assignConnection(weakConnection);
         Event.createOrReuse(createdEvent);
         // TODO: handle stopid, startid
         return createdEvent;
@@ -62,7 +56,6 @@ public class ConnectionEvents {
         } else {
             // event was trashed
             Event trashedEvent = JsonConverter.retrieveEventFromJson(json);
-            trashedEvent.assignConnection(weakConnection);
             trashedEvent.setId(eventId);
             Event.createOrReuse(trashedEvent);
             return eventId;
@@ -72,7 +65,6 @@ public class ConnectionEvents {
     public Event update(Event updateEvent) throws IOException {
         HttpClient.ApiResponse apiResponse = httpClient.updateRequest(PATH, updateEvent.getId(), updateEvent).exec();
         Event updatedEvent = JsonConverter.retrieveEventFromJson(apiResponse.getJsonBody());
-        updatedEvent.assignConnection(weakConnection);
         updatedEvent.setId(updateEvent.getId());
         Event.createOrReuse(updatedEvent);
         return updatedEvent;
