+++ commons/src/main/java/com/pryv/connection/ConnectionStreams.java
@@ -1,25 +1,30 @@
 package com.pryv.connection;
 
 
-import com.pryv.AbstractConnection;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.pryv.Filter;
 import com.pryv.api.HttpClient;
 import com.pryv.model.Stream;
 import com.pryv.utils.JsonConverter;
 
 import java.io.IOException;
-import java.lang.ref.WeakReference;
+import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 public class ConnectionStreams {
 
-    private WeakReference<AbstractConnection> weakConnection;
     private HttpClient httpClient;
     private static final String PATH = "streams";
 
-    public ConnectionStreams(WeakReference<AbstractConnection> weakConnection, HttpClient client) {
-        this.weakConnection = weakConnection;
+    private Map<String, Stream> rootStreams;
+    private Map<String, Stream> flatStreams;
+
+    public ConnectionStreams(HttpClient client) {
         this.httpClient = client;
+        this.rootStreams = new ConcurrentHashMap<>();
+        // TODO: Init streams structure with an initial get call + use of recomputeRootStreamsTree
+        this.flatStreams = new ConcurrentHashMap<>();
     }
 
     public Map<String, Stream> get(Filter filter) throws IOException {
@@ -27,20 +32,16 @@ public class ConnectionStreams {
         String json = apiResponse.getJsonBody();
         Map<String, Stream> receivedStreams =
                 JsonConverter.createStreamsTreeFromJson(json);
-        for (Stream receivedStream : receivedStreams.values()) {
-            receivedStream.assignConnection(weakConnection);
-        }
         // TODO: retrieve streamDeletions
         Map<String, Double> streamDeletions =
                 JsonConverter.createStreamDeletionsTreeFromJson(json);
-        weakConnection.get().updateRootStreams(receivedStreams);
+        rootStreams = receivedStreams;
         return receivedStreams;
     }
 
     public Stream create(Stream newStream) throws IOException {
         HttpClient.ApiResponse apiResponse = httpClient.createRequest(PATH, newStream, null).exec();
         Stream createdStream = JsonConverter.retrieveStreamFromJson(apiResponse.getJsonBody());
-        createdStream.assignConnection(weakConnection);
         return createdStream;
     }
 
@@ -54,7 +55,6 @@ public class ConnectionStreams {
         } else {
             // stream was trashed
             Stream trashedStream = JsonConverter.retrieveStreamFromJson(json);
-            trashedStream.assignConnection(weakConnection);
             return trashedStream.getId();
         }
     }
@@ -62,8 +62,42 @@ public class ConnectionStreams {
     public Stream update(Stream streamToUpdate) throws IOException {
         HttpClient.ApiResponse apiResponse = httpClient.updateRequest(PATH, streamToUpdate.getId(), streamToUpdate).exec();
         Stream updatedStream = JsonConverter.retrieveStreamFromJson(apiResponse.getJsonBody());
-        updatedStream.assignConnection(weakConnection);
         return updatedStream;
+    }
+
+    public Map<String, Stream> getRootStreams() {
+        return rootStreams;
+    }
+
+    /**
+     * fixes Streams' children properties based on parentIds
+     */
+    private void recomputeRootStreamsTree() {
+        rootStreams.clear();
+
+        String parentId = null;
+        // set root streams
+        for (Stream potentialRootStream : flatStreams.values()) {
+            // clear children fields
+            potentialRootStream.clearChildren();
+            parentId = potentialRootStream.getParentId();
+            if (parentId == null) {
+                rootStreams.put(potentialRootStream.getId(), potentialRootStream);
+            }
+        }
+
+        // assign children
+        for (Stream childStream : flatStreams.values()) {
+            parentId = childStream.getParentId();
+            if (parentId != null) {
+                if (flatStreams.containsKey(parentId)) {
+                    Stream parent = flatStreams.get(parentId);
+                    if (parent != null) {
+                        parent.addChildStream(childStream);
+                    }
+                }
+            }
+        }
     }
 
 }
